requirements.txt
pytest
selenium
requests
allure-pytest
allure-python-commons